move and slightly refactor some code
